software/service-thrustmaster/requirements.txt
inputs==0.5
python-engineio==3.13.2
python-socketio==4.6.1
python-socketio-client==1.1